+++ src/tokens.ts
@@ -3,6 +3,7 @@ import { DocumentRegistry } from '@jupyterlab/docregistry';
 import { Token, JSONObject } from '@lumino/coreutils';
 import { IDisposable } from '@lumino/disposable';
 import { ISignal } from '@lumino/signaling';
+import { CommandRegistry } from '@lumino/commands';
 
 export const EXTENSION_ID = 'jupyter.extensions.git_plugin';
 
@@ -56,6 +57,8 @@ export interface IGitExtension extends IDisposable {
    * A signal emitted when the current status of the git repository changes.
    */
   readonly statusChanged: ISignal<IGitExtension, Git.IStatusFileResult[]>;
+
+  readonly commands: CommandRegistry | null;
 
   /**
    * Make request to add one or all files into
